feat(shop-item): correctly pulling shop-item data and displaying
